Moved FactDb to core package
