+++ app/src/test/java/com/veselovvv/numberfactstestapp/data/facts/FactsDbToDataMapperTest.kt
@@ -1,5 +1,6 @@
 package com.veselovvv.numberfactstestapp.data.facts
 
+import com.veselovvv.numberfactstestapp.data.core.FactDb
 import org.junit.Assert.assertEquals
 import org.junit.Test
